Revert "Revert "refactor: move room audio controls to their own component""
This reverts commit 219e0670ca2c0c5e7bb1c25d4928cdc787934c09.
