+++ src/components/Room/Room.tsx
@@ -1,18 +1,9 @@
-import { useState } from 'react'
 import Accordion from '@mui/material/Accordion'
 import AccordionSummary from '@mui/material/AccordionSummary'
 import AccordionDetails from '@mui/material/AccordionDetails'
 import Box from '@mui/material/Box'
 import Divider from '@mui/material/Divider'
 import ExpandMoreIcon from '@mui/icons-material/ExpandMore'
-import RecordVoiceOver from '@mui/icons-material/RecordVoiceOver'
-import VoiceOverOff from '@mui/icons-material/VoiceOverOff'
-import List from '@mui/material/List'
-import ListItem from '@mui/material/ListItem'
-import ListItemText from '@mui/material/ListItemText'
-import Menu from '@mui/material/Menu'
-import MenuItem from '@mui/material/MenuItem'
-import Fab from '@mui/material/Fab'
 import { v4 as uuid } from 'uuid'
 
 import { rtcConfig } from 'config/rtcConfig'
@@ -21,6 +12,7 @@ import { MessageForm } from 'components/MessageForm'
 import { ChatTranscript } from 'components/ChatTranscript'
 
 import { useRoom } from './useRoom'
+import { RoomAudioControls } from './RoomAudioControls'
 
 export interface RoomProps {
   appId?: string
@@ -37,15 +29,7 @@ export function Room({
   password,
   userId,
 }: RoomProps) {
-  const {
-    audioDevices,
-    messageLog,
-    sendMessage,
-    isMessageSending,
-    isSpeakingToRoom,
-    setIsSpeakingToRoom,
-    handleAudioDeviceSelect,
-  } = useRoom(
+  const { messageLog, peerRoom, sendMessage, isMessageSending } = useRoom(
     {
       appId,
       trackerUrls,
@@ -59,35 +43,8 @@ export function Room({
     }
   )
 
-  const [audioAnchorEl, setAudioAnchorEl] = useState<null | HTMLElement>(null)
-  const isAudioDeviceSelectOpen = Boolean(audioAnchorEl)
-  const [selectedAudioDeviceIdx, setSelectedAudioDeviceIdx] = useState(0)
-
   const handleMessageSubmit = async (message: string) => {
     await sendMessage(message)
-  }
-
-  const handleVoiceCallClick = () => {
-    setIsSpeakingToRoom(!isSpeakingToRoom)
-  }
-
-  const handleAudioDeviceListItemClick = (
-    event: React.MouseEvent<HTMLElement>
-  ) => {
-    setAudioAnchorEl(event.currentTarget)
-  }
-
-  const handleAudioDeviceMenuItemClick = (
-    _event: React.MouseEvent<HTMLElement>,
-    idx: number
-  ) => {
-    setSelectedAudioDeviceIdx(idx)
-    handleAudioDeviceSelect(audioDevices[idx])
-    setAudioAnchorEl(null)
-  }
-
-  const handleAudioInputSelectMenuClose = () => {
-    setAudioAnchorEl(null)
   }
 
   return (
@@ -114,70 +71,7 @@ export function Room({
               justifyContent: 'center',
             }}
           >
-            <Fab
-              variant="extended"
-              color={isSpeakingToRoom ? 'error' : 'success'}
-              aria-label="call"
-              onClick={handleVoiceCallClick}
-            >
-              {isSpeakingToRoom ? (
-                <>
-                  <VoiceOverOff sx={{ mr: 1 }} />
-                  Stop speaking to room
-                </>
-              ) : (
-                <>
-                  <RecordVoiceOver sx={{ mr: 1 }} />
-                  Start speaking to room
-                </>
-              )}
-            </Fab>
-            {audioDevices.length > 0 && (
-              <Box sx={{ mt: 1 }}>
-                <List
-                  component="nav"
-                  aria-label="Audio device selection"
-                  sx={{ bgcolor: 'background.paper' }}
-                >
-                  <ListItem
-                    button
-                    id="audio-input-select-button"
-                    aria-haspopup="listbox"
-                    aria-controls="audio-input-select-menu"
-                    aria-label="Audio input device to use"
-                    aria-expanded={isAudioDeviceSelectOpen ? 'true' : undefined}
-                    onClick={handleAudioDeviceListItemClick}
-                  >
-                    <ListItemText
-                      primary="Selected audio input device"
-                      secondary={audioDevices[selectedAudioDeviceIdx]?.label}
-                    />
-                  </ListItem>
-                </List>
-                <Menu
-                  id="audio-input-select-menu"
-                  anchorEl={audioAnchorEl}
-                  open={isAudioDeviceSelectOpen}
-                  onClose={handleAudioInputSelectMenuClose}
-                  MenuListProps={{
-                    'aria-labelledby': 'audio-input-select-button',
-                    role: 'listbox',
-                  }}
-                >
-                  {audioDevices.map((audioDevice, idx) => (
-                    <MenuItem
-                      key={audioDevice.deviceId}
-                      selected={idx === selectedAudioDeviceIdx}
-                      onClick={event =>
-                        handleAudioDeviceMenuItemClick(event, idx)
-                      }
-                    >
-                      {audioDevice.label}
-                    </MenuItem>
-                  ))}
-                </Menu>
-              </Box>
-            )}
+            <RoomAudioControls peerRoom={peerRoom} />
           </Box>
         </AccordionDetails>
       </Accordion>
